Fix for support post width with small nozzles
Since I ask people to lie and use a small nozzle width if they want models to exactly match lego bricks, it's not okay to have things depend *only* on the nozzle width. I did a review and found the support posts on 2xN bricks failed that check. This is now fixed.